requirements.txt
python-dotenv==1.0.1
coloredlogs==15.0.1